PIM_System-master/requirement.txt
boto==2.49.0
boto3==1.9.145
botocore==1.12.145
Django==2.1.7
django-bootstrap4==0.0.8
django-storages==1.7.1
djangorestframework==3.9.3
docutils==0.14
jmespath==0.9.4
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.3
